+++ static/js/server_events_dispatch.js
@@ -330,30 +330,41 @@ exports.dispatch_normal_event = function dispatch_normal_event(event) {
                 }
             }
         } else if (event.op === 'peer_add') {
-            function add_peer(stream_name, user_id) {
-                if (!stream_data.add_subscriber(stream_name, user_id)) {
+            function add_peer(stream_id, user_id) {
+                const sub = stream_data.get_sub_by_id(stream_id);
+
+                if (!sub) {
+                    blueslip.warn('Cannot find stream for peer_add: ' + stream_id);
+                    return;
+                }
+
+                if (!stream_data.add_subscriber(sub.name, user_id)) {
                     blueslip.warn('Cannot process peer_add event');
                     return;
                 }
 
-                stream_edit.rerender(stream_name);
+                stream_edit.rerender(sub.name);
+                compose_fade.update_faded_users();
             }
-            for (const stream_name of event.subscriptions) {
-                add_peer(stream_name, event.user_id);
-            }
-            compose_fade.update_faded_users();
+            add_peer(event.stream_id, event.user_id);
         } else if (event.op === 'peer_remove') {
-            function remove_peer(stream_name, user_id) {
-                if (!stream_data.remove_subscriber(stream_name, user_id)) {
+            function remove_peer(stream_id, user_id) {
+                const sub = stream_data.get_sub_by_id(stream_id);
+
+                if (!sub) {
+                    blueslip.warn('Cannot find stream for peer_remove: ' + stream_id);
+                    return;
+                }
+
+                if (!stream_data.remove_subscriber(sub.name, user_id)) {
                     blueslip.warn('Cannot process peer_remove event.');
                     return;
                 }
-                stream_edit.rerender(stream_name);
+
+                stream_edit.rerender(sub.name);
+                compose_fade.update_faded_users();
             }
-            for (const stream_name of event.subscriptions) {
-                remove_peer(stream_name, event.user_id);
-            }
-            compose_fade.update_faded_users();
+            remove_peer(event.stream_id, event.user_id);
         } else if (event.op === 'remove') {
             for (const rec of event.subscriptions) {
                 const sub = stream_data.get_sub_by_id(rec.stream_id);